Add a pea record and de-anonymise lambdas.
The pea record will group together the various components and should be
the thing that gets passed around to most of the sub-systems. This is in
preparation for adding the player module.

The anonymous functions used in libev watchers have been moved into
function generators.

This helps me track variable dependencies.

And also improves indent levels a little..