Server send VFS vpath = false when vpath unchanged.
This makes it easier for clients to determine when to request a list of tracks.

ie, a VFS message can now indicate that only a cursor position has changed, rather than a vpath change.

Previously, clients would have to maintain state to check when a vpath has changed, or brute-force always request tracks even when only the cursor position has changed.