Change read-trim-right to read only one trailing char.
This is because input-port-ready? might only be for the Chez scheme internal port buffer.

There's an edge case where if a scheme datum, when as represented in a bytevector, has a length of 1024 (the default custom port buffer size), then there's no further reading past the last datum char.

But that leaves a trailing newline, that is ignored because 'read' has just read all it needs to, so there's no accidental extra chars sitting in the port buffer and pea is left blocking on read or waiting in the event loop for new packet arrivals.

I suspect the proper generic solution be to implement non-blocking socket ports.

But in the meantime, this pea-specific-protocol-method is safe to use if all pea client/server instances append one newline to each sent datum (as is done transparently by 'write-now').